custom_commands/requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
cryptography==36.0.1
Deprecated==1.2.13
h11==0.13.0
idna==3.3
outcome==1.1.0
pycparser==2.21
pyfiglet==0.8.post1
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
requests==2.27.1
selenium==4.1.2
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
termcolor==1.1.0
trio==0.20.0
trio-websocket==0.9.2
typing==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.8
webdriver-manager==3.5.3
wrapt==1.14.1
wsproto==1.1.0
